feat: implement Excel to PDF conversion configuration and processing logic
