+++ procedure_generator/requirements.txt
@@ -6,5 +6,7 @@ python-docx==1.2.0
 six==1.17.0
 playwright>=1.52.0
 PyYAML>=6.0
-pydantic>=2.0.0
-pydantic-settings>=2.0.0
+pydantic==2.11.7
+pydantic-settings==2.10.0
+pandas==2.3.0
+openpyxl==3.1.5
